lots of new linting (stricter new ruff)

+++ matchms/filtering/peak_processing/require_minimum_number_of_peaks.py
@@ -43,7 +43,7 @@ def _require_minimum_number_of_peaks_spectrum(
 
     parent_mass = spectrum.get("parent_mass", None)
     if parent_mass and ratio_required:
-        n_required_by_mass = int(ceil(ratio_required * parent_mass))
+        n_required_by_mass = ceil(ratio_required * parent_mass)
         threshold = max(n_required, n_required_by_mass)
     else:
         threshold = n_required
